+++ zakurad/src/components/inbound/tests/real_peer_set.rs
@@ -832,7 +832,7 @@ async fn setup(
 
     // Inbound
     let (setup_tx, setup_rx) = oneshot::channel();
-    let inbound_service = Inbound::new(MAX_INBOUND_CONCURRENCY, setup_rx);
+    let inbound_service = Inbound::new(MAX_INBOUND_CONCURRENCY, false, setup_rx);
     // TODO: add a timeout just above the service, if needed
     let inbound_service = ServiceBuilder::new()
         .load_shed()
@@ -902,6 +902,7 @@ async fn setup(
     let mempool_config = MempoolConfig::default();
     let (mut mempool_service, transaction_subscriber) = Mempool::new(
         &mempool_config,
+        false,
         peer_set.clone(),
         state_service.clone(),
         buffered_tx_verifier.clone(),
@@ -1060,7 +1061,7 @@ mod submitblock_test {
 
         // Inbound
         let (_setup_tx, setup_rx) = oneshot::channel();
-        let inbound_service = Inbound::new(MAX_INBOUND_CONCURRENCY, setup_rx);
+        let inbound_service = Inbound::new(MAX_INBOUND_CONCURRENCY, false, setup_rx);
         let inbound_service = ServiceBuilder::new()
             .load_shed()
             .buffer(10)